src/requirements.txt
cryptography>=3.4
python-dotenv>=0.19